+++ galsim/__init__.py
@@ -131,6 +131,7 @@ from .compound import AutoConvolve, AutoConvolution, AutoCorrelate, AutoCorrelat
 from .compound import FourierSqrt, FourierSqrtProfile
 from .compound import RandomWalk
 from .transform import Transform, Transformation, _Transform
+from .vonkarman import VonKarman
 
 # Chromatic
 from .chromatic import ChromaticObject, ChromaticAtmosphere, ChromaticSum