+++ CHANGELOG.md
@@ -17,7 +17,7 @@ Bug Fixes
 - Fixed a number of small bugs in the config processing uncovered by the
   galsim_extra FocalPlane output type. (#928)
 - Fixed python3 unicode/str mismatches in tests/SConscript (#932)
-
+- Fixed memory leak when drawing PhaseScreenPSFs using photon-shooting (#942)
 
 Deprecated Features
 -------------------
@@ -27,6 +27,7 @@ Deprecated Features
 New Features
 ------------
 - Add option to use circular weight function in HSM adaptive moments code. (#917)
+- Add VonKarman profile GSObject.